Todo funcionando, solo faltan los filtros de comidas

+++ Aplicación/Mod_Ingrediente/TiposIngredientes/Panaderia.cs
@@ -8,7 +8,7 @@ namespace Aplicacion
 {
     public class Panaderia : PorPeso
     {
-        public Panaderia(int codigo, string nombre, int cantidadminima, decimal precioKilo, decimal cantidad)
+        public Panaderia(int codigo, string nombre, decimal cantidadminima, decimal precioKilo, decimal cantidad)
         {
             Codigo = codigo;
             Nombre = nombre;
